+++ frontend_app/android/app/src/main/kotlin/com/aiqinban/app/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.aiqinban.app
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()